requirements/base.txt
pytz==2022.7.1  # https://github.com/stub42/pytz
python-slugify==7.0.0  # https://github.com/un33k/python-slugify
Pillow==9.4.0  # https://github.com/python-pillow/Pillow
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.3.0  # https://github.com/evansd/whitenoise
redis==4.4.2  # https://github.com/redis/redis-py
hiredis==2.1.1  # https://github.com/redis/hiredis-py
sorl-thumbnail==12.9.0  # https://github.com/jazzband/sorl-thumbnail
sorl-thumbnail-serializer-field==0.2.1  # https://github.com/dessibelle/sorl-thumbnail-serializer-field
django-ordered-model==3.6  # https://github.com/django-ordered-model/django-ordered-model

# Django
# ------------------------------------------------------------------------------
django==4.0.8  # pyup: < 4.1  # https://www.djangoproject.com/
django-environ==0.9.0  # https://github.com/joke2k/django-environ
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.52.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.14.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.7  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.2.0  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.14.0  # https://github.com/encode/django-rest-framework
django-cors-headers==3.13.0  # https://github.com/adamchainz/django-cors-headers
# DRF-spectacular for api documentation
drf-spectacular==0.25.1  # https://github.com/tfranzel/drf-spectacular
django-filter==22.1
djangorestframework-camel-case==1.3.0
